added maps for d10_1038_s41467_018_06318_7 d10_1038_s41586_019_1373_2 d10_1038_s41586_019_1652_y
