+++ sfaira/data/dataloaders/loaders/d10_1038_s41586_019_1652_y/human_liver_2019_10x_popescu_001.py
@@ -15,12 +15,12 @@ class Dataset(DatasetBase):
             **kwargs
     ):
         super().__init__(data_path=data_path, meta_path=meta_path, cache_path=cache_path, **kwargs)
-        self.id = "human_liver_2019_10x_popescu_001_10.1038/s41586-019-1652-y"
+        self.id = "human_liver_2019_10xsequencing_popescu_001_10.1038/s41586-019-1652-y"
 
         self.download_url_data = "private,fetal_liver_alladata_.h5ad"
         self.download_url_meta = None
 
-        self.author = "Haniffa"
+        self.author = "Popescu"
         self.doi = "10.1038/s41586-019-1652-y"
         self.healthy = True
         self.normalization = "raw"
@@ -33,38 +33,6 @@ class Dataset(DatasetBase):
         self.var_symbol_col = "index"
 
         self.obs_key_cellontology_original = "cell.labels"
-
-        self.class_maps = {
-            "0": {
-                "B cell": "Mature B cells",
-                "DC1": "Dendritic cell 1",
-                "DC2": "Dendritic cell 2",
-                "DC precursor": "Dendritic cell precursor",
-                "Early Erythroid": "Early Erythroid",
-                "Early lymphoid_T lymphocyte": "Early lymphoid T lymphocyte",
-                "Endothelial cell": "Endothelial cell",
-                "Fibroblast": "Fibroblast",
-                "HSC_MPP": "HSC MPP",
-                "Hepatocyte": "Hepatocyte",
-                "ILC precursor": "ILC precursor",
-                "Kupffer Cell": "Kupffer Cell",
-                "Late Erythroid": "Late Erythroid",
-                "MEMP": "MEMP",
-                "Mast cell": "Mast cell",
-                "Megakaryocyte": "Megakaryocyte",
-                "Mid Erythroid": "Mid Erythroid",
-                "Mono-Mac": "Mono Macrophage",
-                "Monocyte": "Monocyte",
-                "Monocyte precursor": "Monocyte precursor",
-                "NK": "NK cell",
-                "Neutrophil-myeloid progenitor": "Neutrophil myeloid progenitor",
-                "Pre pro B cell": "Pre pro B cell",
-                "VCAM1+ EI macrophage": "VCAM1pos EI macrophage",
-                "pDC precursor": "pDendritic cell precursor",
-                "pre-B cell": "pre B cell",
-                "pro-B cell": "pro B cell"
-            },
-        }
 
     def _load(self):
         fn = os.path.join(self.data_dir, "fetal_liver_alladata_.h5ad")
